third_party/blink/web_tests/fast/events/popup-blocking-timers5-expected.txt

Click Here
PASS successfullyParsed is true

TEST COMPLETE

Test calling window.open() with a 900 ms delay. A popup should be allowed.
PASS newWindow is non-null.
